refactor(grpcapi): split server_diag.go by concern (#4658)
server_diag.go had grown to 1602 LOC, folding four unrelated concerns
into one review unit: streamed diagnostic exec commands, long-lived
streaming interface/packet-drop monitors, the destructive SystemAction /
peer-proxy verb dispatch, and the security-critical factory-reset zeroize
suite (#4582/#4595/#4601). Reviewing a wipe path in the same file as a
1-second monitor ticker is error-prone; this splits the file into
cohesive siblings so each concern is a self-contained review unit.

This is PURE code motion — no behavior change. Every top-level
declaration is moved verbatim (bytes, doc comments, and all streaming
cancellation / WaitDelay / io.Pipe-close / peer-proxy recursion metadata
preserved), and the destructive zeroize helpers are byte-identical since
they are wipe paths that must not drift.

New sibling files in package grpcapi:

  - server_diag_ping.go        Ping, Traceroute, their argv builders, and
                               the shared streamDiagCmd exec-streamer.
  - server_diag_monitor.go     MonitorPacketDrop, MonitorInterface,
                               proxyMonitorInterface, the RuntimeDataPlane
                               adapter glue, the node/interface-alias
                               validators, and monitorSummaryModeFromProto.
  - server_diag_system_action.go  SystemAction verb dispatch,
                               proxyPeerSystemAction, logSystemAction,
                               schedulePowerAction.
  - server_diag_zeroize.go     The factory-reset wipe suite: zeroizeConfigDir,
                               zeroizeRenderedConfigs, zeroizeLoginAccounts,
                               performZeroizeWipe, and their helpers/vars/consts.

server_diag.go now holds only dialPeer — the peer-fabric dial primitive
shared by both the monitor proxy and the system-action proxy.

Validation:
  - go build ./... clean; go vet ./pkg/grpcapi/... clean; gofmt clean.
  - go test ./pkg/grpcapi/ green, including the security-critical suite:
    TestZeroize* (13 cases), TestStreamDiagCmd*, TestSystemActionJournals*,
    TestIsTextRollbackFile, TestClampDiagTimeout.
  - AST byte-identity check: all 30 top-level declarations byte-for-byte
    identical between origin/master's server_diag.go and the union of the
    new files (0 bodydiff / 0 missing / 0 extra; identical count 30==30).

Closes #4658
