+++ pkg/grpcapi/server_diag.go
@@ -1,655 +1,20 @@
 package grpcapi
 
 import (
-	"bufio"
 	"context"
-	"errors"
 	"fmt"
-	"io"
 	"log/slog"
 	"net"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"sort"
-	"strconv"
-	"strings"
 	"syscall"
 	"time"
 
-	"github.com/psaab/xpf/pkg/appid"
-	"github.com/psaab/xpf/pkg/cluster"
-	"github.com/psaab/xpf/pkg/config"
-	dpuserspace "github.com/psaab/xpf/pkg/dataplane/userspace"
-	dpformat "github.com/psaab/xpf/pkg/dataplane/userspace/format"
-	"github.com/psaab/xpf/pkg/dhcpserver"
-	"github.com/psaab/xpf/pkg/diagcmd"
-	"github.com/psaab/xpf/pkg/frr"
 	pb "github.com/psaab/xpf/pkg/grpcapi/xpfv1"
-	"github.com/psaab/xpf/pkg/ipsec"
-	"github.com/psaab/xpf/pkg/monitoriface"
 	"golang.org/x/sys/unix"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials/insecure"
-	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
 )
-
-var _ monitoriface.RuntimeDataPlane = monitorInterfaceServerDataPlane{}
-
-type monitorInterfaceServerDataPlane struct {
-	server *Server
-}
-
-func (a monitorInterfaceServerDataPlane) IsLoaded() bool {
-	return a.server != nil && a.server.dp != nil && a.server.dp.IsLoaded()
-}
-
-func (a monitorInterfaceServerDataPlane) ReadInterfaceCounters(ifindex int) (monitoriface.InterfaceCounters, error) {
-	if a.server == nil || a.server.dp == nil {
-		return monitoriface.InterfaceCounters{}, fmt.Errorf("dataplane unavailable")
-	}
-	ctrs, err := a.server.dp.ReadInterfaceCounters(ifindex)
-	if err != nil {
-		return monitoriface.InterfaceCounters{}, err
-	}
-	return monitoriface.InterfaceCounters{
-		RxPackets: ctrs.RxPackets,
-		RxBytes:   ctrs.RxBytes,
-		TxPackets: ctrs.TxPackets,
-		TxBytes:   ctrs.TxBytes,
-	}, nil
-}
-
-func (s *Server) monitorInterfaceDataplane() monitoriface.RuntimeDataPlane {
-	if s == nil || s.dp == nil {
-		return nil
-	}
-	return monitorInterfaceServerDataPlane{server: s}
-}
-
-// --- Diagnostic RPCs ---
-
-func (s *Server) Ping(req *pb.PingRequest, stream grpc.ServerStreamingServer[pb.PingResponse]) error {
-	if req.Target == "" {
-		return status.Error(codes.InvalidArgument, "target required")
-	}
-	count := int(req.Count)
-	if count <= 0 {
-		count = 5
-	}
-	if count > 100 {
-		count = 100
-	}
-	cmd := buildPingArgv(req, count)
-
-	return streamDiagCmd(stream.Context(), pingExecTimeout(count), cmd, func(line string) error {
-		return stream.Send(&pb.PingResponse{Output: line})
-	})
-}
-
-// buildPingArgv builds the argv for the gRPC Ping diag. It delegates to
-// the shared diagcmd builder so the VRF-device normalization (apply
-// "vrf-" exactly once, #2143) and the "--" end-of-options separator
-// (option-confusion hardening, #2084) match the CLI and REST surfaces
-// byte-for-byte. count is the already-clamped probe count.
-func buildPingArgv(req *pb.PingRequest, count int) []string {
-	size := ""
-	if req.Size > 0 {
-		size = fmt.Sprintf("%d", req.Size)
-	}
-	return diagcmd.PingArgv(diagcmd.PingOptions{
-		Target:          req.Target,
-		Count:           fmt.Sprintf("%d", count),
-		Source:          req.Source,
-		Size:            size,
-		RoutingInstance: req.RoutingInstance,
-	})
-}
-
-func (s *Server) Traceroute(req *pb.TracerouteRequest, stream grpc.ServerStreamingServer[pb.TracerouteResponse]) error {
-	if req.Target == "" {
-		return status.Error(codes.InvalidArgument, "target required")
-	}
-	cmd := buildTracerouteArgv(req)
-
-	return streamDiagCmd(stream.Context(), diagTracerouteTimeout, cmd, func(line string) error {
-		return stream.Send(&pb.TracerouteResponse{Output: line})
-	})
-}
-
-// buildTracerouteArgv builds the argv for the gRPC Traceroute diag.
-// Like buildPingArgv it delegates to the shared diagcmd builder so VRF
-// normalization (#2143) and the "--" separator (#2084) stay identical
-// across the CLI, REST, and gRPC surfaces.
-func buildTracerouteArgv(req *pb.TracerouteRequest) []string {
-	return diagcmd.TracerouteArgv(diagcmd.TracerouteOptions{
-		Target:          req.Target,
-		Source:          req.Source,
-		RoutingInstance: req.RoutingInstance,
-	})
-}
-
-// streamDiagCmd runs a command and streams each line of combined output
-// via sendFn. timeout is request-sized by the caller (#1819, see the
-// diag-stream budget block in exec_timeout.go); it is always capped at
-// diagExecCeiling so a pathological request cannot pin the handler.
-func streamDiagCmd(ctx context.Context, timeout time.Duration, cmd []string, sendFn func(string) error) error {
-	ctx, cancelTimeout := context.WithTimeout(ctx, clampDiagTimeout(timeout))
-	defer cancelTimeout()
-	// Cancelable layer for the send-failure path: when sendFn fails the
-	// scanner goroutine stops reading the pipe, so without an explicit
-	// cancel the child would block on writes until the deadline killed
-	// it — cancel here makes CommandContext kill it promptly (#1819).
-	ctx, cancel := context.WithCancel(ctx)
-	defer cancel()
-	c := exec.CommandContext(ctx, cmd[0], cmd[1:]...)
-	// U3 parity (#1805): cap the post-kill pipe-drain window so a child
-	// that inherited the output pipe cannot block Wait past the ctx kill.
-	c.WaitDelay = requestExecWaitDelay
-
-	// Merge stdout and stderr into a single pipe.
-	pr, pw := io.Pipe()
-	c.Stdout = pw
-	c.Stderr = pw
-
-	if err := c.Start(); err != nil {
-		return status.Errorf(codes.Internal, "exec: %v", err)
-	}
-
-	// Scan lines and stream each one.
-	scanner := bufio.NewScanner(pr)
-	scanDone := make(chan error, 1)
-	go func() {
-		for scanner.Scan() {
-			if err := sendFn(scanner.Text()); err != nil {
-				// Nobody reads the pipe after this point — kill the
-				// child now instead of letting it block on writes for
-				// the rest of the budget, AND close the read end:
-				// exec.Cmd's internal copy goroutine may already be
-				// blocked in pw.Write on a burst the scanner never
-				// consumed, and WaitDelay only closes the exec-owned
-				// OS pipes, not this io.Pipe — pr.Close makes that
-				// blocked write return ErrClosedPipe so c.Wait can
-				// finish (Codex review on PR #1823).
-				cancel()
-				pr.Close()
-				scanDone <- err
-				return
-			}
-		}
-		scanDone <- scanner.Err()
-	}()
-
-	// Wait for the command to finish, then close the write end so scanner terminates.
-	cmdErr := c.Wait()
-	pw.Close()
-	scanErr := <-scanDone
-
-	if scanErr != nil {
-		return scanErr
-	}
-	if cmdErr != nil {
-		// Send the exit error as a final line rather than failing the RPC,
-		// so the client still sees partial output (e.g. "ping: unknown host").
-		_ = sendFn(cmdErr.Error())
-	}
-	return nil
-}
-
-// MonitorPacketDrop streams packet drop events matching the request filters.
-func (s *Server) MonitorPacketDrop(req *pb.MonitorPacketDropRequest, stream grpc.ServerStreamingServer[pb.MonitorPacketDropResponse]) error {
-	if s.eventBuf == nil {
-		return status.Error(codes.Unavailable, "event buffer not available")
-	}
-
-	// Validate the request before subscribing (#3382). MonitorPacketDrop
-	// streams only LOCAL drops; the CLI applies the same guardrails. An
-	// unvalidated filter silently produces an empty stream that looks like
-	// "no drops" during incident response, so reject impossible inputs up
-	// front with InvalidArgument rather than opening a stream that can never
-	// match.
-
-	// node: packet-drop is local-only — it does not proxy to the cluster
-	// peer the way MonitorInterface does. Honor an empty/local node and
-	// reject any value that designates a different (or "all") node so a
-	// client cannot believe a peer filter was applied.
-	if !s.isLocalNodeRef(req.Node) {
-		return status.Errorf(codes.InvalidArgument,
-			"node %q: monitor packet-drop is local-only; run it on the target node", req.Node)
-	}
-
-	// count: negative is not "unlimited" (the CLI rejects it); enforce the
-	// CLI's 1..8192 cap. 0 remains an explicit "unlimited" sentinel.
-	if req.Count < 0 {
-		return status.Errorf(codes.InvalidArgument, "count must be >= 0 (0 = unlimited), got %d", req.Count)
-	}
-	if req.Count > 8192 {
-		return status.Errorf(codes.InvalidArgument, "count must be 1..8192 (0 = unlimited), got %d", req.Count)
-	}
-
-	// ports: the event records carry 16-bit ports; a value > 65535 can
-	// never match and would silently empty the stream.
-	if req.SourcePort > 65535 {
-		return status.Errorf(codes.InvalidArgument, "source-port must be 0..65535, got %d", req.SourcePort)
-	}
-	if req.DestinationPort > 65535 {
-		return status.Errorf(codes.InvalidArgument, "destination-port must be 0..65535, got %d", req.DestinationPort)
-	}
-
-	// protocol: validate against the shared catalog (case-insensitive,
-	// named or numeric). A typo like "tpc" can never match. Resolve the
-	// request to its protocol NUMBER once so the filter loop can compare
-	// numerically — the event record renders rec.Protocol as a NAME for the
-	// named set, so a numeric request like "6" must not be string-compared
-	// against "TCP" (the accepted-but-never-matches bug).
-	var reqProtoNum uint8
-	if req.Protocol != "" {
-		n, ok := appid.ProtocolNumber(req.Protocol)
-		if !ok {
-			return status.Errorf(codes.InvalidArgument, "unknown protocol %q", req.Protocol)
-		}
-		reqProtoNum = n
-	}
-
-	// zone / interface: validate against the active configuration so a typo
-	// (`trsut`, `ge-0/0/99`) is rejected rather than emitting nothing. The
-	// interface filter must MATCH on whichever alias the daemon stored, so
-	// resolve the full alias set once and reuse it in the loop.
-	var reqIfaceAliases map[string]bool
-	if req.FromZone != "" || req.Interface != "" {
-		cfg := s.store.ActiveConfig()
-		if cfg == nil {
-			return status.Error(codes.Unavailable, "no active configuration to validate zone/interface filters")
-		}
-		if req.FromZone != "" {
-			if _, ok := cfg.Security.Zones[req.FromZone]; !ok {
-				return status.Errorf(codes.InvalidArgument, "unknown from-zone %q", req.FromZone)
-			}
-		}
-		if req.Interface != "" {
-			reqIfaceAliases = interfaceAliasSet(cfg, req.Interface)
-			if len(reqIfaceAliases) == 0 {
-				return status.Errorf(codes.InvalidArgument, "unknown interface %q", req.Interface)
-			}
-		}
-	}
-
-	// Parse filters.
-	var srcNet, dstNet *net.IPNet
-	if req.SourcePrefix != "" {
-		_, cidr, err := net.ParseCIDR(req.SourcePrefix)
-		if err != nil {
-			ip := net.ParseIP(req.SourcePrefix)
-			if ip == nil {
-				return status.Errorf(codes.InvalidArgument, "invalid source-prefix: %s", req.SourcePrefix)
-			}
-			if ip.To4() != nil {
-				cidr = &net.IPNet{IP: ip, Mask: net.CIDRMask(32, 32)}
-			} else {
-				cidr = &net.IPNet{IP: ip, Mask: net.CIDRMask(128, 128)}
-			}
-		}
-		srcNet = cidr
-	}
-	if req.DestinationPrefix != "" {
-		_, cidr, err := net.ParseCIDR(req.DestinationPrefix)
-		if err != nil {
-			ip := net.ParseIP(req.DestinationPrefix)
-			if ip == nil {
-				return status.Errorf(codes.InvalidArgument, "invalid destination-prefix: %s", req.DestinationPrefix)
-			}
-			if ip.To4() != nil {
-				cidr = &net.IPNet{IP: ip, Mask: net.CIDRMask(32, 32)}
-			} else {
-				cidr = &net.IPNet{IP: ip, Mask: net.CIDRMask(128, 128)}
-			}
-		}
-		dstNet = cidr
-	}
-
-	count := int(req.Count)
-	sub := s.eventBuf.Subscribe(256)
-	defer sub.Close()
-
-	if err := stream.Send(&pb.MonitorPacketDropResponse{Line: "Starting packet drop:"}); err != nil {
-		return err
-	}
-
-	seen := 0
-	ctx := stream.Context()
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case rec := <-sub.C:
-			if rec.Type != "POLICY_DENY" && rec.Type != "SCREEN_DROP" {
-				continue
-			}
-			if srcNet != nil {
-				host, _, _ := net.SplitHostPort(rec.SrcAddr)
-				if host == "" {
-					host = rec.SrcAddr
-				}
-				if ip := net.ParseIP(host); ip == nil || !srcNet.Contains(ip) {
-					continue
-				}
-			}
-			if dstNet != nil {
-				host, _, _ := net.SplitHostPort(rec.DstAddr)
-				if host == "" {
-					host = rec.DstAddr
-				}
-				if ip := net.ParseIP(host); ip == nil || !dstNet.Contains(ip) {
-					continue
-				}
-			}
-			if req.SourcePort != 0 {
-				_, portStr, _ := net.SplitHostPort(rec.SrcAddr)
-				if p, _ := strconv.ParseUint(portStr, 10, 16); p != uint64(req.SourcePort) {
-					continue
-				}
-			}
-			if req.DestinationPort != 0 {
-				_, portStr, _ := net.SplitHostPort(rec.DstAddr)
-				if p, _ := strconv.ParseUint(portStr, 10, 16); p != uint64(req.DestinationPort) {
-					continue
-				}
-			}
-			if req.Protocol != "" && rec.ProtocolNum != reqProtoNum {
-				// Compare against the numeric protocol the RECORD carries,
-				// not a re-parse of the rendered name. Comparing numerically
-				// is robust regardless of the name SSOT: even though #3393
-				// closed the ProtocolName↔ProtocolNumber round-trip (proto 41
-				// "ipv6" now reverses), the raw number (#3382) keeps the match
-				// total for named, numeric-only, and any future render-only
-				// protocols alike. reqProtoNum is resolved once in validation
-				// via appid.ProtocolNumber (correct for the named/numeric
-				// REQUEST side).
-				continue
-			}
-			if req.FromZone != "" && rec.InZoneName != req.FromZone {
-				continue
-			}
-			if req.Interface != "" && !reqIfaceAliases[rec.IngressIface] {
-				// rec.IngressIface is the single daemon-resolved name; the
-				// request may be any accepted alias (config key, Linux
-				// form, or Name override). Match against the full alias set
-				// so any accepted form matches (#3382 matcher fix).
-				continue
-			}
-
-			// Format the drop event.
-			ts := rec.Time.Format("15:04:05.000000")
-			reason := rec.Action
-			if rec.ScreenCheck != "" {
-				reason = "Dropped by SCREEN:" + rec.ScreenCheck
-			} else if rec.Type == "POLICY_DENY" {
-				reason = "Dropped by FLOW:Policy deny"
-				if rec.PolicyName != "" {
-					reason = "Dropped by FLOW:Policy " + rec.PolicyName
-				}
-			}
-			line := fmt.Sprintf("%s %s-->%s;%s,%s,%s",
-				ts, rec.SrcAddr, rec.DstAddr,
-				strings.ToLower(rec.Protocol),
-				rec.IngressIface, reason)
-
-			if err := stream.Send(&pb.MonitorPacketDropResponse{Line: line}); err != nil {
-				return err
-			}
-			seen++
-			if count > 0 && seen >= count {
-				return nil
-			}
-		}
-	}
-}
-
-// isLocalNodeRef reports whether a request's node reference designates the
-// local node (#3382). An empty string or "local" is always local. In a
-// cluster the node's numeric id ("0"/"1") or "nodeN" form for THIS node is
-// accepted; a standalone daemon (localID 0) accepts "0" or "node0". Anything
-// else — the peer id, "all", "primary" — is not local, so a local-only RPC
-// must reject it rather than silently filter local data.
-func (s *Server) isLocalNodeRef(node string) bool {
-	switch strings.ToLower(strings.TrimSpace(node)) {
-	case "", "local":
-		return true
-	}
-	ref := strings.ToLower(strings.TrimSpace(node))
-	localID := 0
-	if s != nil && s.cluster != nil {
-		localID = s.cluster.NodeID()
-	}
-	return ref == strconv.Itoa(localID) || ref == fmt.Sprintf("node%d", localID)
-}
-
-// interfaceAliasSet returns the set of equivalent names for the configured
-// interface that `name` designates, or nil if it matches none (#3382). The
-// packet-drop filter compares against the SINGLE daemon-resolved ingress name
-// (resolveIfName), but the operator may pass any accepted alias — the config
-// key ("ge-0/0/1"), its Linux-form rendering ("ge-0-0-1"), an explicit Name
-// override, or that Name's Linux form. Returning the full alias set lets both
-// validation (len > 0) and the matcher (membership test) accept whichever form
-// the daemon happened to store, so a validated interface filter cannot
-// accept-but-never-match.
-func interfaceAliasSet(cfg *config.Config, name string) map[string]bool {
-	if cfg == nil {
-		return nil
-	}
-	for key, ifc := range cfg.Interfaces.Interfaces {
-		aliases := map[string]bool{
-			key:                     true,
-			config.LinuxIfName(key): true,
-		}
-		if ifc != nil && ifc.Name != "" {
-			aliases[ifc.Name] = true
-			aliases[config.LinuxIfName(ifc.Name)] = true
-		}
-		if aliases[name] {
-			return aliases
-		}
-	}
-	return nil
-}
-
-// MonitorInterface streams pre-formatted interface statistics frames.
-func (s *Server) MonitorInterface(req *pb.MonitorInterfaceRequest, stream grpc.ServerStreamingServer[pb.MonitorInterfaceResponse]) error {
-	cfg := s.store.ActiveConfig()
-	if cfg == nil {
-		return status.Error(codes.Unavailable, "no active configuration")
-	}
-
-	hostname, _ := os.Hostname()
-	if hostname == "" {
-		hostname = "xpf"
-	}
-
-	// resolveToKernel converts a config-level interface name to its kernel name.
-	// e.g. "ge-0/0/0" → "ge-0-0-0", "reth0" → physical member's kernel name.
-	resolveToKernel := func(cfgName string) string {
-		resolved := cfg.ResolveReth(cfgName)
-		return config.LinuxIfName(resolved)
-	}
-
-	// isRethName returns true if the name is a RETH interface (reth0, reth1, etc).
-	isRethName := func(name string) bool {
-		return strings.HasPrefix(name, "reth")
-	}
-
-	// rethRG returns the redundancy group for a RETH interface, or -1.
-	rethRG := func(name string) int {
-		parts := strings.SplitN(name, ".", 2)
-		if ifc, ok := cfg.Interfaces.Interfaces[parts[0]]; ok && ifc.RedundancyGroup > 0 {
-			return ifc.RedundancyGroup
-		}
-		return -1
-	}
-
-	// isPeerInterface returns true if the named interface is a cluster peer's
-	// physical member. Checks FPC slot → node-id mapping first (ge-7/0/X on
-	// node0 belongs to node1), then falls back to RG interface monitors.
-	isPeerInterface := func(name string) bool {
-		if cfg.Chassis.Cluster == nil {
-			return false
-		}
-		// Check FPC slot: slot 0 → node0, slot 7 → node1.
-		if s.cluster != nil {
-			slot := config.InterfaceSlot(name)
-			if slot >= 0 && config.SlotToNodeID(slot) != s.cluster.NodeID() {
-				return true
-			}
-		}
-		base := strings.SplitN(name, ".", 2)[0]
-		// Check RG interface monitors.
-		for _, rg := range cfg.Chassis.Cluster.RedundancyGroups {
-			for _, mon := range rg.InterfaceMonitors {
-				if strings.SplitN(mon.Interface, ".", 2)[0] == base {
-					return true
-				}
-			}
-		}
-		return false
-	}
-
-	isSingle := req.InterfaceName != ""
-	var singleDisplayName, singleKernelName string
-	if isSingle {
-		singleDisplayName = req.InterfaceName
-		singleKernelName = monitoriface.ResolvePhysicalParent(resolveToKernel(req.InterfaceName))
-
-		// Check if interface should be proxied to the cluster peer.
-		needProxy := false
-		if _, err := net.InterfaceByName(singleKernelName); err != nil {
-			// Interface doesn't exist locally. Check if it's a peer's physical member.
-			if isPeerInterface(req.InterfaceName) {
-				needProxy = true
-			} else {
-				return status.Errorf(codes.NotFound, "interface %s not found", req.InterfaceName)
-			}
-		} else if isRethName(req.InterfaceName) {
-			// RETH exists locally but may be MASTER on the peer node.
-			if rg := rethRG(req.InterfaceName); rg > 0 && s.cluster != nil && !s.cluster.IsLocalPrimary(rg) {
-				needProxy = true
-			}
-		}
-
-		if needProxy {
-			return s.proxyMonitorInterface(req, stream)
-		}
-	}
-
-	summaryInterfaces := func() ([]string, map[string]string) {
-		if names, kernelNames := monitoriface.TrafficSummaryInterfaces(cfg); len(names) > 0 {
-			return names, kernelNames
-		}
-
-		c := s.store.ActiveConfig()
-		if c == nil || c.Interfaces.Interfaces == nil {
-			return nil, nil
-		}
-		names := make([]string, 0, len(c.Interfaces.Interfaces))
-		kernelNames := make(map[string]string, len(c.Interfaces.Interfaces))
-		for name := range c.Interfaces.Interfaces {
-			names = append(names, name)
-			kernelNames[name] = monitoriface.ResolvePhysicalParent(resolveToKernel(name))
-		}
-		sort.Strings(names)
-		return names, kernelNames
-	}
-
-	startTime := time.Now()
-	ctx := stream.Context()
-	ticker := time.NewTicker(time.Second)
-	defer ticker.Stop()
-	summaryMode := monitorSummaryModeFromProto(req.GetSummaryMode())
-
-	// Previous snapshots for rate calculation.
-	var prevSingle *monitoriface.Snapshot
-	var baselineSingle *monitoriface.Snapshot
-	prevAll := make(map[string]*monitoriface.Snapshot)
-
-	readSnap := func(name string) *monitoriface.Snapshot {
-		snap, err := monitoriface.ReadSnapshot(s.monitorInterfaceDataplane(), s.userspaceDataplaneStatus, name)
-		if err != nil {
-			return nil
-		}
-		return &snap
-	}
-
-	for {
-		var buf strings.Builder
-		if isSingle {
-			snap := readSnap(singleKernelName)
-			if snap == nil {
-				fmt.Fprintf(&buf, "interface %s: not available\n", singleDisplayName)
-			} else {
-				if baselineSingle == nil {
-					baselineSingle = snap
-				}
-				monitoriface.RenderSingleInterface(&buf, hostname, singleDisplayName, singleKernelName, snap, prevSingle, baselineSingle, startTime)
-				snapCopy := *snap
-				prevSingle = &snapCopy
-			}
-		} else {
-			names, kernelNames := summaryInterfaces()
-			snaps := make(map[string]*monitoriface.Snapshot, len(names))
-			newPrev := make(map[string]*monitoriface.Snapshot, len(names))
-			for _, name := range names {
-				snap := readSnap(kernelNames[name])
-				if snap == nil {
-					continue
-				}
-				newPrev[name] = snap
-				snaps[name] = snap
-			}
-			monitoriface.RenderTrafficSummary(&buf, hostname, names, kernelNames, snaps, prevAll, summaryMode, startTime)
-			prevAll = newPrev
-		}
-
-		if err := stream.Send(&pb.MonitorInterfaceResponse{Frame: buf.String()}); err != nil {
-			return err
-		}
-
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		case <-ticker.C:
-		}
-	}
-}
-
-// proxyMonitorInterface forwards a MonitorInterface stream to the cluster peer.
-func (s *Server) proxyMonitorInterface(req *pb.MonitorInterfaceRequest, stream grpc.ServerStreamingServer[pb.MonitorInterfaceResponse]) error {
-	conn, err := s.dialPeer()
-	if err != nil {
-		return err
-	}
-	defer conn.Close()
-
-	ctx := stream.Context()
-
-	client := pb.NewBpfrxServiceClient(conn)
-	peerStream, err := client.MonitorInterface(ctx, req)
-	if err != nil {
-		return status.Errorf(codes.Unavailable, "peer monitor failed: %v", err)
-	}
-
-	for {
-		resp, err := peerStream.Recv()
-		if err != nil {
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
-			return err
-		}
-		if err := stream.Send(resp); err != nil {
-			return err
-		}
-	}
-}
 
 // dialPeer establishes a gRPC connection to the cluster peer via the fabric link.
 // Tries fab0 first, then fab1 if dual-fabric is configured and fab0 fails.
@@ -709,894 +74,4 @@ func (s *Server) dialPeer() (*grpc.ClientConn, error) {
 		return conn, nil
 	}
 	return nil, status.Errorf(codes.Unavailable, "cannot connect to peer on any fabric address: %v", lastErr)
-}
-
-func (s *Server) proxyPeerSystemAction(ctx context.Context, req *pb.SystemActionRequest) (*pb.SystemActionResponse, error) {
-	peerCtx, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-	peerCtx = metadata.AppendToOutgoingContext(peerCtx, "x-peer-forwarded", "1")
-	if s.peerSystemActionFn != nil {
-		return s.peerSystemActionFn(peerCtx, req)
-	}
-	conn, err := s.dialPeer()
-	if err != nil {
-		return nil, err
-	}
-	defer conn.Close()
-	return pb.NewBpfrxServiceClient(conn).SystemAction(peerCtx, req)
-}
-
-// logSystemAction records a destructive maintenance action to the configstore
-// audit journal BEFORE it executes (#4108 F8). The store append is synchronous
-// and fsynced, so an attributable record is durable on disk before the action
-// runs — the slog.Warn line only reaches journald, which does not survive a
-// reboot. For reboot/halt/power-off the on-disk journal record persists across
-// the reboot; for zeroize the wipe now deliberately removes .config.journal
-// (#4576 — a completed factory reset must not hand its audit log to the next
-// tenant), so the durable cross-wipe trail is the pre-execution fsync (an
-// interrupted wipe still leaves it) plus any remote syslog collector.
-// Best-effort: a journal write failure never blocks a confirmed power/wipe
-// action (the store layer warns rather than failing).
-func (s *Server) logSystemAction(action string) {
-	if s.store == nil {
-		return
-	}
-	s.store.LogSystemAction(action)
-}
-
-// schedulePowerAction runs `systemctl <arg>` after a 1s grace so the RPC
-// response reaches the client first. It is a package var so a test can drive
-// the reboot/halt/power-off SystemAction verbs (to assert the #4108 F8 journal
-// wiring) WITHOUT actually taking the host down.
-var schedulePowerAction = func(systemctlArg string) {
-	go func() {
-		time.Sleep(1 * time.Second)
-		// context.Background(): a confirmed power action must not be
-		// cancelled by client disconnect. Errors ignored as before.
-		runTimeout(context.Background(), "systemctl", systemctlArg)
-	}()
-}
-
-// defaultConfigDir / defaultConfigBase mirror the daemon's config-path default
-// (cmd/xpfd `-config /etc/xpf/xpf.conf`, pkg/daemon). performZeroizeWipe erases
-// the fixed appliance paths; a non-default `-config` location is out of scope
-// (the pre-#4576 wipe already assumed /etc/xpf).
-const (
-	defaultConfigDir  = "/etc/xpf"
-	defaultConfigBase = "xpf.conf"
-)
-
-// zeroizeConfigDir erases the xpf configuration STATE under configDir as part
-// of a factory reset (#4576): the .configdb SSOT + master.key, the numbered
-// text rollback slots, the top-level .conf files, and the audit journal — the
-// artifacts that persist the prior tenant's committed policy, IKE PSKs,
-// WireGuard private keys, and SNMP communities and would otherwise be reloaded
-// on the next boot.
-//
-// NOTE (scope): this erases the SSOT and rollback/journal state under
-// configDir. The RENDERED service configs xpfd writes OUTSIDE configDir —
-// /etc/frr/frr.conf (0644, BGP-MD5/OSPF/ISIS auth), /etc/swanctl/conf.d/xpf.conf
-// (IKE PSKs), /etc/kea/kea-dhcp{4,6}.conf — are erased separately by
-// zeroizeRenderedConfigs (#4585). Both run under performZeroizeWipe. Erasing
-// the rendered configs in the wipe (rather than deferring to the reboot) is
-// required because a post-zeroize boot has no committed config and SKIPS the
-// reconcile that would otherwise clear them — see zeroizeRenderedConfigs.
-//
-// The artifacts removed (configBase is the config file's base name, e.g.
-// "xpf.conf", used to recognize the numbered text rollback slots):
-//   - .configdb/master.key      — the AES-GCM key that decrypts an encrypted DB
-//   - .configdb/                — the SSOT: active.json, candidate.json,
-//     rollback.N.json (Store.Load reloads active.json on next boot)
-//   - <configBase>.<N>          — the CANONICAL text rollback slots
-//     (saveRollbackFiles / loadRollbackHistory), full config TEXT with
-//     cleartext secret leaves; loadRollbackHistory reloads them at boot, so
-//     leaving them behind would let the prior tenant's config be rolled back
-//     to — the exact leak #4576 closes
-//   - *.conf                    — the live config + rescue.conf (pre-#4576 set)
-//   - rollback*                 — legacy rollback naming (pre-#4576 set)
-//   - .config.journal[.N]       — the JSONL audit journal + rotated segments
-//     (0644, prior-tenant commit history/comments; legacy v1 fat lines may
-//     carry full config incl. secrets). This SUPERSEDES the #4108 "journal
-//     survives a zeroize" belt-and-braces: the system_action record is still
-//     fsynced BEFORE the wipe (so an interrupted wipe leaves a trail, and a
-//     remote syslog collector keeps the durable cross-wipe record), but a
-//     completed factory reset must not hand its audit log to the next owner.
-//   - tls/                      — the self-signed REST-API TLS pair (#4599):
-//     tls/key.pem (the device-generated localhost HTTPS private key, 0600) +
-//     tls/cert.pem. xpf-generated, not tenant config; generateSelfSignedCertAt
-//     (pkg/api) regenerates a fresh pair on absence at the next boot, so
-//     removing them is safe and hands no prior-tenant key to the next owner.
-//
-// Removal is KEY-FIRST: master.key is deleted before the encrypted DB body, so
-// an interrupted wipe (crash / power loss mid-RemoveAll) can never leave the
-// ciphertext behind together with the key that decrypts it — once the key is
-// gone any surviving ciphertext is unrecoverable.
-//
-// It is best-effort past a failure (a single stubborn file does not abort the
-// rest of the erasure) but returns the FIRST real error so a
-// silently-incomplete zeroize is never reported as a successful factory reset.
-// os.ErrNotExist is not an error here (an already-absent artifact is the goal).
-func zeroizeConfigDir(configDir, configBase string) error {
-	var firstErr error
-	fail := func(err error) {
-		if err != nil && !errors.Is(err, os.ErrNotExist) && firstErr == nil {
-			firstErr = err
-		}
-	}
-
-	dbDir := filepath.Join(configDir, ".configdb")
-	// KEY-FIRST (#4576): master.key before the encrypted DB body.
-	fail(os.Remove(filepath.Join(dbDir, "master.key")))
-	// The config SSOT (active.json, candidate.json, rollback.N.json + any
-	// residual key). RemoveAll erases the whole tree and is nil on absent.
-	fail(os.RemoveAll(dbDir))
-
-	// The self-signed REST-API TLS material (#4599): tls/key.pem (the
-	// device-generated localhost HTTPS private key) + tls/cert.pem. These are
-	// xpf-generated, not tenant config — generateSelfSignedCertAt (pkg/api)
-	// regenerates a fresh pair on absence at the next boot, so removing them is
-	// safe. A subdir, so the top-level ReadDir loop's os.Remove never catches it;
-	// erase the whole tree explicitly.
-	fail(os.RemoveAll(filepath.Join(configDir, "tls")))
-
-	// Top-level artifacts in a single ReadDir pass.
-	entries, err := os.ReadDir(configDir)
-	fail(err)
-	for _, f := range entries {
-		name := f.Name()
-		if strings.HasSuffix(name, ".conf") ||
-			strings.HasPrefix(name, "rollback") ||
-			name == ".config.journal" ||
-			strings.HasPrefix(name, ".config.journal.") ||
-			isTextRollbackFile(name, configBase) {
-			fail(os.Remove(filepath.Join(configDir, name)))
-		}
-	}
-	return firstErr
-}
-
-// isTextRollbackFile reports whether name is a numbered text rollback slot for
-// the config file configBase — "<configBase>.<N>" with N one-or-more decimal
-// digits (store_commit.go rollbackPath). These files carry the full prior
-// config text including cleartext secret leaves, so a factory reset removes
-// them. configBase itself ("xpf.conf") is caught by the .conf suffix rule.
-func isTextRollbackFile(name, configBase string) bool {
-	rest, ok := strings.CutPrefix(name, configBase+".")
-	if !ok || rest == "" {
-		return false
-	}
-	for _, r := range rest {
-		if r < '0' || r > '9' {
-			return false
-		}
-	}
-	return true
-}
-
-// zeroizeRenderedConfigs erases the RENDERED service configs xpfd writes
-// OUTSIDE configDir — the artifacts that hold the prior tenant's secrets in
-// cleartext but are not part of the .configdb SSOT wiped by zeroizeConfigDir
-// (#4585, follow-up to #4576):
-//
-//   - frrConf (/etc/frr/frr.conf, mode 0644 WORLD-READABLE): the xpf-managed
-//     section carries BGP MD5 / OSPF / IS-IS authentication keys. Only that
-//     section is stripped (frr.StripManagedSectionFile) — operator content
-//     outside the markers is preserved and a purely operator-managed frr.conf
-//     is left untouched. No FRR reload: FRR restarts clean on the reboot.
-//   - swanctlSnippet (/etc/swanctl/conf.d/xpf.conf): the IKE PSKs live in this
-//     single xpf-owned snippet, so it is removed outright.
-//   - kea4/kea6 (/etc/kea/kea-dhcp{4,6}.conf): xpf owns these whole files, so
-//     they are removed outright.
-//
-// WHY the wipe must erase these DIRECTLY rather than lean on the completing
-// reboot: a post-zeroize boot has NO committed config, so the daemon enters
-// #1922 bootstrap mode (or, on an HA node, a normal boot with a nil active
-// config) and SKIPS the boot-time applyConfig that would otherwise reconcile
-// FRR/IPsec/Kea to empty (pkg/daemon/daemon_run.go: bootstrap suppresses the
-// apply, and the normal-boot apply is gated on ActiveConfig() != nil). The
-// rendered secrets are therefore a PERSISTENT residual across the reboot, not a
-// transient one — a device handed to the next tenant would keep the prior
-// tenant's routing-auth keys in a world-readable file. (See docs/system-login.md.)
-//
-// Discipline mirrors zeroizeConfigDir: os.ErrNotExist is not an error (an
-// already-absent artifact is the goal), removal is best-effort past a single
-// failure, but the FIRST real error is returned so a silently-incomplete wipe
-// is never reported as a clean factory reset.
-func zeroizeRenderedConfigs(frrConf, swanctlSnippet, kea4, kea6 string) error {
-	var firstErr error
-	fail := func(err error) {
-		if err != nil && !errors.Is(err, os.ErrNotExist) && firstErr == nil {
-			firstErr = err
-		}
-	}
-	// FRR: strip only the xpf-managed section (the routing-auth secrets); the
-	// file may carry operator content outside the markers. StripManagedSectionFile
-	// already treats an absent file as a no-op (nil) and leaves an unmanaged
-	// frr.conf untouched.
-	fail(frr.StripManagedSectionFile(frrConf))
-	// swanctl snippet + Kea configs: xpf owns these whole files, remove them.
-	fail(os.Remove(swanctlSnippet))
-	fail(os.Remove(kea4))
-	fail(os.Remove(kea6))
-	return firstErr
-}
-
-// zeroize login-account teardown paths (#4598). These mirror the production
-// paths the daemon provisions OS login accounts under (pkg/daemon:
-// provisionedUsersDir, sudoersDir/sudoersPrefix, passwdPath, /home/<user>).
-// They cannot be imported — pkg/daemon imports pkg/grpcapi (daemon_run.go),
-// so a shared symbol would create an import cycle, the same reason the exec
-// timeout helper is duplicated here. They are package vars only so a test can
-// point the teardown at a throwaway tree instead of the real /etc + /home +
-// /var/lib.
-var (
-	// zeroizeProvisionedUsersDir is the per-account provenance-marker directory
-	// (pkg/daemon.provisionedUsersDir). Each file is named for an xpf-created
-	// login account and CONTAINS that account's UID at provision time — the
-	// UID-keyed ownership marker of #1944. It is the authoritative registry of
-	// "which OS users did xpf provision", so the wipe walks it to know exactly
-	// which accounts are safe to remove.
-	zeroizeProvisionedUsersDir = "/var/lib/xpf/provisioned-users"
-	// zeroizeSudoersDir + zeroizeSudoersPrefix mirror pkg/daemon.sudoersDir /
-	// sudoersPrefix — the exclusive namespace of xpf-managed NOPASSWD sudo
-	// drop-ins. Only files with the prefix are ever removed; operator-authored
-	// drop-ins in the same directory are left untouched.
-	zeroizeSudoersDir    = "/etc/sudoers.d"
-	zeroizeSudoersPrefix = "xpf-"
-	// zeroizePasswdPath is /etc/passwd, read to resolve an account's CURRENT
-	// UID so a userdel only ever fires when the live UID still matches the
-	// marker (never on an out-of-band recreate). Mirrors pkg/daemon.passwdPath.
-	zeroizePasswdPath = "/etc/passwd"
-	// zeroizeHomeBase is the home-directory root; authorized_keys lives at
-	// <base>/<user>/.ssh/authorized_keys.
-	zeroizeHomeBase = "/home"
-	// zeroizeUserdel removes an OS account, its /etc/shadow + /etc/passwd
-	// entries, and its home tree (-r). It is a seam so tests can record the
-	// removal without touching real accounts. context.Background(): a client
-	// disconnect must not abort a confirmed factory reset (mirrors runTimeout's
-	// power-action rationale).
-	zeroizeUserdel = func(name string) ([]byte, error) {
-		return combinedOutputTimeout(context.Background(), "userdel", "-r", name)
-	}
-)
-
-// zeroizeLookupUID returns the current numeric UID for name by parsing
-// zeroizePasswdPath directly (cgo-free, mirroring pkg/daemon.lookupUID). It
-// returns (uid, true) on success and (0, false) if the file is unreadable, the
-// account is absent, or the UID field does not parse — the "not our account"
-// disposition that suppresses a userdel.
-func zeroizeLookupUID(name string) (int, bool) {
-	data, err := os.ReadFile(zeroizePasswdPath)
-	if err != nil {
-		return 0, false
-	}
-	for _, line := range strings.Split(string(data), "\n") {
-		if line == "" {
-			continue
-		}
-		fields := strings.Split(line, ":")
-		if len(fields) < 3 {
-			continue
-		}
-		if fields[0] == name {
-			uid, err := strconv.Atoi(fields[2])
-			if err != nil {
-				return 0, false
-			}
-			return uid, true
-		}
-	}
-	return 0, false
-}
-
-// zeroizeLoginAccounts tears down the OS LOGIN accounts xpf provisioned as part
-// of a factory reset (#4598, follow-up to #4576/#4585). These are the biggest
-// re-tenant leak of the three because they grant INTERACTIVE LOGIN + SUDO, not
-// just a config-secret read:
-//
-//   - /etc/shadow password hashes (pkg/daemon.reconcileUserPassword)
-//   - SSH authorized_keys under /home/<user>/.ssh (pkg/daemon.applySystemLogin)
-//   - /etc/sudoers.d/xpf-<user> NOPASSWD grants (pkg/daemon.reconcileSudoers)
-//
-// All three live OUTSIDE /etc/xpf and SURVIVE the config wipe: applySystemLogin
-// runs only inside the boot-time applyConfig, which a post-zeroize boot SKIPS
-// (bootstrap mode / nil active config — the same boot-skip #4585 documents), and
-// even a full reconcile early-returns on an empty config and never userdel's. So
-// without this teardown a re-tenanted / RMA'd / resold device still carries the
-// prior tenant's login accounts, SSH keys, and passwordless sudo.
-//
-// Ownership discipline — NEVER touch a non-xpf account (the core safety
-// invariant, #1944 §5.4):
-//   - Sudoers: only the xpf-<user> namespace is swept. Operator-authored
-//     drop-ins (no xpf- prefix) are left untouched. The whole namespace is
-//     removed unconditionally — at factory reset nothing is desired — so no
-//     passwordless-root grant survives even if a marker is missing.
-//   - Users: a userdel fires ONLY for an account that has a provenance marker
-//     AND whose CURRENT /etc/passwd UID still equals the marker's recorded UID.
-//     An operator's own account (no marker) is never iterated; a system account
-//     or root (no marker) is never touched; an out-of-band userdel+recreate with
-//     a different UID (marker mismatch) is left alone — exactly the #1944
-//     leave-then-rejoin vs recreate distinction, so the wipe can never nuke the
-//     wrong account or strand access.
-//
-// authorized_keys is removed BEFORE userdel so the SSH-key login vector dies
-// even if userdel later fails; the marker is retained on a userdel FAILURE so a
-// retried zeroize re-attempts the removal (and the failure is surfaced, so the
-// device is not reported safe to re-tenant while a live account remains).
-//
-// Discipline mirrors zeroizeConfigDir / zeroizeRenderedConfigs: os.ErrNotExist
-// is never an error (an already-absent artifact is the goal), removal is
-// best-effort past a single failure, but the FIRST real error is returned so a
-// silently-incomplete teardown is never reported as a clean factory reset.
-func zeroizeLoginAccounts() error {
-	var firstErr error
-	fail := func(err error) {
-		if err != nil && !errors.Is(err, os.ErrNotExist) && firstErr == nil {
-			firstErr = err
-		}
-	}
-
-	// (A) Sweep the xpf sudoers namespace. Removing every xpf-<user> drop-in
-	// guarantees no passwordless-root grant survives, independent of the marker
-	// registry. Operator drop-ins without the prefix are never touched.
-	if entries, err := os.ReadDir(zeroizeSudoersDir); err != nil {
-		fail(err) // a real ReadDir error (absent dir is ErrNotExist → excluded)
-	} else {
-		for _, e := range entries {
-			name := e.Name()
-			if e.IsDir() || !strings.HasPrefix(name, zeroizeSudoersPrefix) {
-				continue
-			}
-			fail(os.Remove(filepath.Join(zeroizeSudoersDir, name)))
-		}
-	}
-
-	// (B) Tear down each xpf-provisioned OS user, gated on the UID-keyed marker.
-	entries, err := os.ReadDir(zeroizeProvisionedUsersDir)
-	if err != nil {
-		// Absent dir = xpf never provisioned a login account: nothing to do.
-		// Any other read error is surfaced (we cannot prove the accounts gone).
-		fail(err)
-		return firstErr
-	}
-	for _, e := range entries {
-		if e.IsDir() {
-			continue
-		}
-		name := e.Name() // marker filename == account name (Base'd on write)
-		markerFile := filepath.Join(zeroizeProvisionedUsersDir, name)
-
-		recordedUID, uidErr := readProvisionedMarkerUID(markerFile)
-		curUID, curOK := zeroizeLookupUID(name)
-		keysFile := filepath.Join(zeroizeHomeBase, name, ".ssh", "authorized_keys")
-
-		removeMarker := true
-		switch {
-		case !curOK:
-			// Account already absent from /etc/passwd — it cannot authenticate.
-			// Best-effort clean up any orphaned key residue; nothing to userdel.
-			fail(os.Remove(keysFile))
-		case uidErr != nil || curUID != recordedUID:
-			// Corrupt marker OR out-of-band recreate (UID changed): NOT the
-			// account xpf provisioned. Never userdel or touch its home — the
-			// current owner is someone else. Drop our stale marker only.
-		default:
-			// curUID == recordedUID → the exact account xpf provisioned. Kill
-			// the SSH-key vector first (survives a userdel failure), then remove
-			// the account (userdel -r drops /etc/shadow + /etc/passwd + home).
-			fail(os.Remove(keysFile))
-			if out, derr := zeroizeUserdel(name); derr != nil {
-				slog.Error("zeroize: failed to remove xpf-provisioned login account",
-					"user", name, "err", derr, "output", strings.TrimSpace(string(out)))
-				fail(derr)
-				removeMarker = false // keep the marker so a retry re-attempts
-			} else {
-				slog.Info("zeroize: removed xpf-provisioned login account", "user", name)
-			}
-		}
-		if removeMarker {
-			fail(os.Remove(markerFile))
-		}
-	}
-	// Drop the now-empty marker directory (best-effort; ENOTEMPTY after a
-	// retained marker is fine and must not gate the factory reset).
-	_ = os.Remove(zeroizeProvisionedUsersDir)
-	return firstErr
-}
-
-// readProvisionedMarkerUID reads a provenance marker file and returns the UID
-// it records (pkg/daemon.markProvisioned writes the account's UID as decimal
-// text). A read or parse error yields the "not our account" disposition in
-// zeroizeLoginAccounts (no userdel).
-func readProvisionedMarkerUID(path string) (int, error) {
-	data, err := os.ReadFile(path)
-	if err != nil {
-		return 0, err
-	}
-	return strconv.Atoi(strings.TrimSpace(string(data)))
-}
-
-// performZeroizeWipe erases the on-disk config, rendered service-config
-// secrets, provisioned login accounts, BPF pins, and managed networkd files
-// (factory reset). It is a package var so a test can drive the `zeroize`
-// SystemAction verb (to assert the #4108 F8 journal wiring) WITHOUT wiping a
-// real /etc/xpf on the developer/appliance box. It returns a non-nil error when
-// a security-critical erasure did not fully complete (#4576/#4585/#4598).
-var performZeroizeWipe = func() error {
-	// Config state FIRST — the security-critical erasure. A failure here can
-	// leave prior-tenant config/secrets on disk, so it is surfaced to the
-	// caller (#4576).
-	err := zeroizeConfigDir(defaultConfigDir, defaultConfigBase)
-
-	// Rendered service configs (#4585): also security-critical — routing-auth
-	// keys in a world-readable frr.conf, IKE PSKs, Kea configs. A post-zeroize
-	// boot enters bootstrap / nil-active-config normal boot and SKIPS the
-	// reconcile that would clear them, so the wipe must erase them itself. Fold
-	// its first error into the surfaced result (the .configdb error takes
-	// priority) so a partial wipe is never reported as a clean factory reset.
-	if e := zeroizeRenderedConfigs(
-		frr.DefaultFRRConf,
-		filepath.Join(ipsec.DefaultSwanctlDir, ipsec.BPFRXConfFile),
-		dhcpserver.DefaultKea4ConfPath,
-		dhcpserver.DefaultKea6ConfPath,
-	); e != nil && err == nil {
-		err = e
-	}
-
-	// Provisioned login accounts (#4598): the OS users xpf created — their
-	// /etc/shadow hashes, SSH authorized_keys, and /etc/sudoers.d/xpf-* grants —
-	// live OUTSIDE /etc/xpf and SURVIVE the config wipe (applySystemLogin runs
-	// only inside the boot-time applyConfig, which a post-zeroize boot SKIPS), so
-	// a re-tenanted device would otherwise grant the prior tenant interactive
-	// login + passwordless sudo. Marker-aware teardown (UID-keyed provenance
-	// marker, #1944) so a non-xpf admin/system/operator account is NEVER touched.
-	// Also security-critical, so fold its first error into the surfaced result.
-	if e := zeroizeLoginAccounts(); e != nil && err == nil {
-		err = e
-	}
-
-	// BPF pins + managed networkd files carry no secret material, so their
-	// removal stays best-effort (logged, never fatal — they do not gate the
-	// success/failure of the factory reset).
-	if e := os.RemoveAll("/sys/fs/bpf/xpf"); e != nil {
-		slog.Warn("zeroize: remove BPF pins failed", "err", e)
-	}
-	if ndFiles, e := os.ReadDir("/etc/systemd/network"); e == nil {
-		for _, f := range ndFiles {
-			if strings.HasPrefix(f.Name(), "10-xpf-") {
-				if re := os.Remove(filepath.Join("/etc/systemd/network", f.Name())); re != nil && !errors.Is(re, os.ErrNotExist) {
-					slog.Warn("zeroize: remove networkd file failed", "file", f.Name(), "err", re)
-				}
-			}
-		}
-	}
-	return err
-}
-
-func (s *Server) SystemAction(ctx context.Context, req *pb.SystemActionRequest) (*pb.SystemActionResponse, error) {
-	switch req.Action {
-	case "reboot":
-		slog.Warn("system reboot requested via gRPC")
-		// Journal BEFORE the box goes down: the fsynced record survives the
-		// reboot even though the slog.Warn journald line may not (#4108 F8).
-		s.logSystemAction("reboot")
-		schedulePowerAction("reboot")
-		return &pb.SystemActionResponse{Message: "System going down for reboot NOW!"}, nil
-
-	case "halt":
-		slog.Warn("system halt requested via gRPC")
-		s.logSystemAction("halt")
-		schedulePowerAction("halt")
-		return &pb.SystemActionResponse{Message: "System halting NOW!"}, nil
-
-	case "power-off":
-		slog.Warn("system power-off requested via gRPC")
-		s.logSystemAction("power-off")
-		schedulePowerAction("poweroff")
-		return &pb.SystemActionResponse{Message: "System powering off NOW!"}, nil
-
-	case "zeroize":
-		slog.Warn("system zeroize requested via gRPC")
-		// Journal BEFORE the wipe: the fsynced record is durable on disk
-		// before any config file is removed, so an interrupted wipe still
-		// leaves an attributable trail and a remote syslog collector keeps the
-		// cross-wipe record (#4108 F8). The wipe itself now removes
-		// .config.journal (a completed factory reset must not hand its audit
-		// log to the next tenant — #4576), superseding the earlier
-		// journal-survives-zeroize belt-and-braces.
-		s.logSystemAction("zeroize")
-		if err := performZeroizeWipe(); err != nil {
-			// A partial wipe may leave prior-tenant config/secrets on disk
-			// (#4576). Surface it instead of reporting a clean factory reset —
-			// the operator must know the erasure did not fully complete and
-			// the device is NOT safe to re-tenant.
-			slog.Error("system zeroize did not fully erase config state", "err", err)
-			return nil, status.Errorf(codes.Internal,
-				"zeroize incomplete: config state may remain on disk: %v", err)
-		}
-		return &pb.SystemActionResponse{Message: "System zeroized. Configuration erased. Reboot to complete factory reset."}, nil
-
-	case "clear-config-lock":
-		holder, locked := s.store.ConfigHolder()
-		if !locked {
-			return &pb.SystemActionResponse{Message: "No configuration lock held"}, nil
-		}
-		s.store.ForceExitConfigure()
-		return &pb.SystemActionResponse{Message: fmt.Sprintf("Configuration lock cleared (was held by %s)", holder)}, nil
-
-	case "clear-arp":
-		out, err := combinedOutputTimeout(ctx, "ip", "-4", "neigh", "flush", "all")
-		if err != nil {
-			return nil, status.Errorf(codes.Internal, "flush ARP: %s", strings.TrimSpace(string(out)))
-		}
-		return &pb.SystemActionResponse{Message: "ARP cache cleared"}, nil
-
-	case "clear-interfaces-statistics":
-		return &pb.SystemActionResponse{
-			Message: "Interface statistics counters noted\n(kernel counters are cumulative and cannot be reset)",
-		}, nil
-
-	case "clear-ipv6-neighbors":
-		out, err := combinedOutputTimeout(ctx, "ip", "-6", "neigh", "flush", "all")
-		if err != nil {
-			return nil, status.Errorf(codes.Internal, "flush IPv6 neighbors: %s", strings.TrimSpace(string(out)))
-		}
-		return &pb.SystemActionResponse{Message: "IPv6 neighbor cache cleared"}, nil
-
-	case "clear-policy-counters":
-		if s.dp == nil || !s.dp.IsLoaded() {
-			return nil, status.Error(codes.Unavailable, "dataplane not loaded")
-		}
-		if err := s.dp.ClearPolicyCounters(); err != nil {
-			return nil, status.Errorf(codes.Internal, "%v", err)
-		}
-		return &pb.SystemActionResponse{Message: "policy hit counters cleared"}, nil
-
-	case "clear-firewall-counters":
-		if s.dp == nil || !s.dp.IsLoaded() {
-			return nil, status.Error(codes.Unavailable, "dataplane not loaded")
-		}
-		if err := s.dp.ClearFilterCounters(); err != nil {
-			return nil, status.Errorf(codes.Internal, "%v", err)
-		}
-		return &pb.SystemActionResponse{Message: "Firewall filter counters cleared"}, nil
-
-	case "clear-nat-counters":
-		if s.dp == nil || !s.dp.IsLoaded() {
-			return nil, status.Error(codes.Unavailable, "dataplane not loaded")
-		}
-		if err := s.dp.ClearNATRuleCounters(); err != nil {
-			return nil, status.Errorf(codes.Internal, "%v", err)
-		}
-		return &pb.SystemActionResponse{Message: "NAT translation statistics cleared"}, nil
-
-	case "clear-persistent-nat":
-		if s.dp == nil || s.dp.GetPersistentNAT() == nil {
-			return &pb.SystemActionResponse{Message: "Persistent NAT table not available"}, nil
-		}
-		count := s.dp.GetPersistentNAT().Len()
-		s.dp.GetPersistentNAT().Clear()
-		return &pb.SystemActionResponse{
-			Message: fmt.Sprintf("Cleared %d persistent NAT bindings", count),
-		}, nil
-
-	case "ospf-clear":
-		if s.frr == nil {
-			return nil, status.Errorf(codes.FailedPrecondition, "FRR manager not available")
-		}
-		if _, err := s.frr.ExecVtysh("clear ip ospf process"); err != nil {
-			return nil, status.Errorf(codes.Internal, "clear OSPF: %v", err)
-		}
-		return &pb.SystemActionResponse{Message: "OSPF process cleared"}, nil
-
-	case "bgp-clear":
-		if s.frr == nil {
-			return nil, status.Errorf(codes.FailedPrecondition, "FRR manager not available")
-		}
-		if _, err := s.frr.ExecVtysh("clear bgp * soft"); err != nil {
-			return nil, status.Errorf(codes.Internal, "clear BGP: %v", err)
-		}
-		return &pb.SystemActionResponse{Message: "BGP sessions cleared (soft reset)"}, nil
-
-	case "ipsec-sa-clear":
-		if s.ipsec == nil {
-			return nil, status.Errorf(codes.FailedPrecondition, "IPsec manager not available")
-		}
-		count, err := s.ipsec.TerminateAllSAs()
-		if err != nil {
-			return nil, status.Errorf(codes.Internal, "%v", err)
-		}
-		return &pb.SystemActionResponse{
-			Message: fmt.Sprintf("Cleared %d IPsec SA(s)", count),
-		}, nil
-
-	case "dhcp-renew":
-		if s.dhcp == nil {
-			return nil, status.Errorf(codes.FailedPrecondition, "DHCP manager not available")
-		}
-		if req.Target == "" {
-			return nil, status.Errorf(codes.InvalidArgument, "dhcp-renew requires target interface")
-		}
-		if err := s.dhcp.Renew(req.Target); err != nil {
-			return nil, status.Errorf(codes.NotFound, "%v", err)
-		}
-		return &pb.SystemActionResponse{
-			Message: fmt.Sprintf("DHCP renewal initiated on %s", req.Target),
-		}, nil
-
-	case "in-service-upgrade":
-		if s.cluster == nil {
-			return nil, status.Error(codes.Unavailable, "cluster not configured")
-		}
-		if err := s.cluster.ForceSecondary(); err != nil {
-			return nil, status.Errorf(codes.FailedPrecondition, "ISSU: %v", err)
-		}
-		return &pb.SystemActionResponse{
-			Message: "Node is now secondary for all redundancy groups.\n" +
-				"Traffic has been drained to peer.\n" +
-				"You may now replace the binary and restart the service:\n" +
-				"  systemctl stop xpfd && <replace binary> && systemctl start xpfd",
-		}, nil
-
-	case "dynamic-dns-update", "dynamic-dns-check":
-		// #3276: operator force-now / check-now verb. force=true (update)
-		// re-asserts every owned DDNS record now; force=false (check) only
-		// re-observes + publishes changed records. The handler honors the
-		// per-RG owner gate (no-op + clear message on the backup).
-		if s.surfaceADDNSForceFn == nil {
-			return nil, status.Error(codes.Unavailable, "dynamic-dns: DDNS engine not running")
-		}
-		_, msg := s.surfaceADDNSForceFn(req.Action == "dynamic-dns-update")
-		return &pb.SystemActionResponse{Message: msg}, nil
-
-	default:
-		// Handle cluster failover actions: "cluster-failover:<rgID>" and "cluster-failover-reset:<rgID>"
-		if strings.HasPrefix(req.Action, "cluster-failover-reset:") {
-			if s.cluster == nil {
-				return nil, status.Error(codes.Unavailable, "cluster not configured")
-			}
-			rgStr := strings.TrimPrefix(req.Action, "cluster-failover-reset:")
-			rgID, err := strconv.Atoi(rgStr)
-			if err != nil {
-				return nil, status.Errorf(codes.InvalidArgument, "invalid redundancy-group ID: %s", rgStr)
-			}
-			if err := s.cluster.ResetFailover(rgID); err != nil {
-				return nil, status.Errorf(codes.NotFound, "%v", err)
-			}
-			return &pb.SystemActionResponse{
-				Message: fmt.Sprintf("Failover reset for redundancy group %d", rgID),
-			}, nil
-		}
-		if strings.HasPrefix(req.Action, "cluster-failover-data:node") {
-			if s.cluster == nil {
-				return nil, status.Error(codes.Unavailable, "cluster not configured")
-			}
-			nodeStr := strings.TrimPrefix(req.Action, "cluster-failover-data:node")
-			targetNode, err := strconv.Atoi(nodeStr)
-			if err != nil {
-				return nil, status.Errorf(codes.InvalidArgument, "invalid node ID: %s", nodeStr)
-			}
-			if !cluster.IsSupportedClusterNodeID(targetNode) {
-				return nil, status.Errorf(codes.InvalidArgument, "unsupported cluster failover target node %d", targetNode)
-			}
-			if targetNode != s.cluster.NodeID() {
-				if peerForwardedFromContext(ctx) {
-					return nil, status.Errorf(codes.FailedPrecondition, "forwarded cluster failover target node %d is not local", targetNode)
-				}
-				resp, err := s.proxyPeerSystemAction(ctx, req)
-				if err != nil {
-					if st, ok := status.FromError(err); ok {
-						return nil, st.Err()
-					}
-					return nil, status.Errorf(codes.Unavailable, "peer cluster failover proxy failed: %v", err)
-				}
-				return resp, nil
-			}
-
-			dataRGs := s.cluster.DataGroupIDs()
-			if len(dataRGs) == 0 {
-				return nil, status.Error(codes.FailedPrecondition, "no data redundancy groups configured")
-			}
-			moveRGs := make([]int, 0, len(dataRGs))
-			for _, rgID := range dataRGs {
-				if !s.cluster.IsLocalPrimary(rgID) {
-					moveRGs = append(moveRGs, rgID)
-				}
-			}
-			if len(moveRGs) == 0 {
-				return &pb.SystemActionResponse{
-					Message: fmt.Sprintf("All data redundancy groups are already primary on node %d", targetNode),
-				}, nil
-			}
-			if len(moveRGs) == 1 {
-				if err := s.cluster.RequestPeerFailover(moveRGs[0]); err != nil {
-					return nil, status.Errorf(codes.FailedPrecondition, "%v", err)
-				}
-			} else {
-				if err := s.cluster.RequestPeerFailoverBatch(moveRGs); err != nil {
-					return nil, status.Errorf(codes.FailedPrecondition, "%v", err)
-				}
-			}
-			return &pb.SystemActionResponse{
-				Message: fmt.Sprintf("Manual failover completed for data redundancy groups %v (transfer committed)", moveRGs),
-			}, nil
-		}
-		if strings.HasPrefix(req.Action, "cluster-failover:") {
-			if s.cluster == nil {
-				return nil, status.Error(codes.Unavailable, "cluster not configured")
-			}
-			rest := strings.TrimPrefix(req.Action, "cluster-failover:")
-
-			// Parse "cluster-failover:<rgID>[:node<N>]"
-			var rgStr, nodeStr string
-			if idx := strings.Index(rest, ":node"); idx >= 0 {
-				rgStr = rest[:idx]
-				nodeStr = rest[idx+5:] // skip ":node"
-			} else {
-				rgStr = rest
-			}
-
-			rgID, err := strconv.Atoi(rgStr)
-			if err != nil {
-				return nil, status.Errorf(codes.InvalidArgument, "invalid redundancy-group ID: %s", rgStr)
-			}
-
-			// If "node <N>" specified, route to correct node.
-			if nodeStr != "" {
-				targetNode, err := strconv.Atoi(nodeStr)
-				if err != nil {
-					return nil, status.Errorf(codes.InvalidArgument, "invalid node ID: %s", nodeStr)
-				}
-				if targetNode != s.cluster.NodeID() {
-					if peerForwardedFromContext(ctx) {
-						return nil, status.Errorf(codes.FailedPrecondition, "forwarded cluster failover target node %d is not local", targetNode)
-					}
-					resp, err := s.proxyPeerSystemAction(ctx, req)
-					if err != nil {
-						if st, ok := status.FromError(err); ok {
-							return nil, st.Err()
-						}
-						return nil, status.Errorf(codes.Unavailable, "peer cluster failover proxy failed: %v", err)
-					}
-					return resp, nil
-				}
-				// Target is local — make us primary.
-				if s.cluster.IsLocalPrimary(rgID) {
-					return &pb.SystemActionResponse{
-						Message: fmt.Sprintf("Redundancy group %d is already primary on node %d", rgID, targetNode),
-					}, nil
-				}
-				// Ask peer to transfer out so we can take primary.
-				if err := s.cluster.RequestPeerFailover(rgID); err != nil {
-					return nil, status.Errorf(codes.FailedPrecondition, "%v", err)
-				}
-				return &pb.SystemActionResponse{
-					Message: fmt.Sprintf("Manual failover completed for redundancy group %d (transfer committed)", rgID),
-				}, nil
-			}
-
-			if err := s.cluster.ManualFailover(rgID); err != nil {
-				return nil, status.Errorf(codes.NotFound, "%v", err)
-			}
-			return &pb.SystemActionResponse{
-				Message: fmt.Sprintf("Manual failover triggered for redundancy group %d", rgID),
-			}, nil
-		}
-		if strings.HasPrefix(req.Action, "userspace-inject:") {
-			provider, err := s.userspaceDataplaneControl()
-			if err != nil {
-				return nil, status.Error(codes.Unavailable, err.Error())
-			}
-			rest := strings.TrimPrefix(req.Action, "userspace-inject:")
-			parts := strings.SplitN(rest, ":", 2)
-			if len(parts) != 2 {
-				return nil, status.Error(codes.InvalidArgument, "usage: userspace-inject:<slot>:<mode>")
-			}
-			slot, err := strconv.Atoi(parts[0])
-			if err != nil {
-				return nil, status.Errorf(codes.InvalidArgument, "invalid userspace slot: %s", parts[0])
-			}
-			mode := parts[1]
-			statusNow, err := provider.Status()
-			if err != nil {
-				return nil, status.Errorf(codes.Unavailable, "userspace status: %v", err)
-			}
-			extra, err := dpuserspace.DecodeInjectPacketTarget(req.Target)
-			if err != nil {
-				return nil, status.Error(codes.InvalidArgument, err.Error())
-			}
-			injectReq, err := dpuserspace.BuildInjectPacketRequest(uint32(slot), mode, extra, statusNow)
-			if err != nil {
-				return nil, status.Error(codes.InvalidArgument, err.Error())
-			}
-			statusAfter, err := provider.InjectPacket(injectReq)
-			if err != nil {
-				return nil, status.Errorf(codes.FailedPrecondition, "userspace inject: %v", err)
-			}
-			msg := dpformat.FormatStatusSummary(statusAfter) + "\n" + dpformat.FormatBindings(statusAfter)
-			return &pb.SystemActionResponse{Message: msg}, nil
-		}
-		if strings.HasPrefix(req.Action, "userspace-forwarding:") {
-			provider, err := s.userspaceDataplaneControl()
-			if err != nil {
-				return nil, status.Error(codes.Unavailable, err.Error())
-			}
-			armed, err := dpuserspace.ParseForwardingCommand([]string{"forwarding", strings.TrimPrefix(req.Action, "userspace-forwarding:")})
-			if err != nil {
-				return nil, status.Error(codes.InvalidArgument, err.Error())
-			}
-			statusAfter, err := provider.SetForwardingArmed(armed)
-			if err != nil {
-				return nil, status.Errorf(codes.FailedPrecondition, "userspace forwarding control: %v", err)
-			}
-			msg := dpformat.FormatStatusSummary(statusAfter) + "\n" + dpformat.FormatBindings(statusAfter)
-			return &pb.SystemActionResponse{Message: msg}, nil
-		}
-		if strings.HasPrefix(req.Action, "userspace-queue:") {
-			provider, err := s.userspaceDataplaneControl()
-			if err != nil {
-				return nil, status.Error(codes.Unavailable, err.Error())
-			}
-			rest := strings.TrimPrefix(req.Action, "userspace-queue:")
-			parts := strings.SplitN(rest, ":", 2)
-			if len(parts) != 2 {
-				return nil, status.Error(codes.InvalidArgument, "usage: userspace-queue:<queue>:<register|unregister|arm|disarm>")
-			}
-			queueID, err := strconv.Atoi(parts[0])
-			if err != nil {
-				return nil, status.Errorf(codes.InvalidArgument, "invalid userspace queue: %s", parts[0])
-			}
-			registered, armed, err := dpuserspace.ParseRegistrationOperation(parts[1])
-			if err != nil {
-				return nil, status.Error(codes.InvalidArgument, err.Error())
-			}
-			statusAfter, err := provider.SetQueueState(uint32(queueID), registered, armed)
-			if err != nil {
-				return nil, status.Errorf(codes.FailedPrecondition, "userspace queue control: %v", err)
-			}
-			msg := dpformat.FormatStatusSummary(statusAfter) + "\n" + dpformat.FormatBindings(statusAfter)
-			return &pb.SystemActionResponse{Message: msg}, nil
-		}
-		if strings.HasPrefix(req.Action, "userspace-binding:") {
-			provider, err := s.userspaceDataplaneControl()
-			if err != nil {
-				return nil, status.Error(codes.Unavailable, err.Error())
-			}
-			rest := strings.TrimPrefix(req.Action, "userspace-binding:")
-			parts := strings.SplitN(rest, ":", 2)
-			if len(parts) != 2 {
-				return nil, status.Error(codes.InvalidArgument, "usage: userspace-binding:<slot>:<register|unregister|arm|disarm>")
-			}
-			slot, err := strconv.Atoi(parts[0])
-			if err != nil {
-				return nil, status.Errorf(codes.InvalidArgument, "invalid userspace slot: %s", parts[0])
-			}
-			registered, armed, err := dpuserspace.ParseRegistrationOperation(parts[1])
-			if err != nil {
-				return nil, status.Error(codes.InvalidArgument, err.Error())
-			}
-			statusAfter, err := provider.SetBindingState(uint32(slot), registered, armed)
-			if err != nil {
-				return nil, status.Errorf(codes.FailedPrecondition, "userspace binding control: %v", err)
-			}
-			msg := dpformat.FormatStatusSummary(statusAfter) + "\n" + dpformat.FormatBindings(statusAfter)
-			return &pb.SystemActionResponse{Message: msg}, nil
-		}
-		return nil, status.Errorf(codes.InvalidArgument, "unknown action: %s", req.Action)
-	}
-}
-
-func monitorSummaryModeFromProto(mode pb.MonitorInterfaceSummaryMode) monitoriface.SummaryMode {
-	switch mode {
-	case pb.MonitorInterfaceSummaryMode_MONITOR_INTERFACE_SUMMARY_MODE_PACKETS:
-		return monitoriface.SummaryModePackets
-	case pb.MonitorInterfaceSummaryMode_MONITOR_INTERFACE_SUMMARY_MODE_BYTES:
-		return monitoriface.SummaryModeBytes
-	case pb.MonitorInterfaceSummaryMode_MONITOR_INTERFACE_SUMMARY_MODE_DELTA:
-		return monitoriface.SummaryModeDelta
-	case pb.MonitorInterfaceSummaryMode_MONITOR_INTERFACE_SUMMARY_MODE_RATE:
-		return monitoriface.SummaryModeRate
-	default:
-		return monitoriface.SummaryModeCombined
-	}
 }
