+++ pkg/grpcapi/server_diag_system_action.go
@@ -1,0 +1,477 @@
+package grpcapi
+
+import (
+	"context"
+	"fmt"
+	"log/slog"
+	"strconv"
+	"strings"
+	"time"
+
+	"github.com/psaab/xpf/pkg/cluster"
+	dpuserspace "github.com/psaab/xpf/pkg/dataplane/userspace"
+	dpformat "github.com/psaab/xpf/pkg/dataplane/userspace/format"
+	pb "github.com/psaab/xpf/pkg/grpcapi/xpfv1"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/metadata"
+	"google.golang.org/grpc/status"
+)
+
+func (s *Server) proxyPeerSystemAction(ctx context.Context, req *pb.SystemActionRequest) (*pb.SystemActionResponse, error) {
+	peerCtx, cancel := context.WithTimeout(ctx, 5*time.Second)
+	defer cancel()
+	peerCtx = metadata.AppendToOutgoingContext(peerCtx, "x-peer-forwarded", "1")
+	if s.peerSystemActionFn != nil {
+		return s.peerSystemActionFn(peerCtx, req)
+	}
+	conn, err := s.dialPeer()
+	if err != nil {
+		return nil, err
+	}
+	defer conn.Close()
+	return pb.NewBpfrxServiceClient(conn).SystemAction(peerCtx, req)
+}
+
+// logSystemAction records a destructive maintenance action to the configstore
+// audit journal BEFORE it executes (#4108 F8). The store append is synchronous
+// and fsynced, so an attributable record is durable on disk before the action
+// runs — the slog.Warn line only reaches journald, which does not survive a
+// reboot. For reboot/halt/power-off the on-disk journal record persists across
+// the reboot; for zeroize the wipe now deliberately removes .config.journal
+// (#4576 — a completed factory reset must not hand its audit log to the next
+// tenant), so the durable cross-wipe trail is the pre-execution fsync (an
+// interrupted wipe still leaves it) plus any remote syslog collector.
+// Best-effort: a journal write failure never blocks a confirmed power/wipe
+// action (the store layer warns rather than failing).
+func (s *Server) logSystemAction(action string) {
+	if s.store == nil {
+		return
+	}
+	s.store.LogSystemAction(action)
+}
+
+// schedulePowerAction runs `systemctl <arg>` after a 1s grace so the RPC
+// response reaches the client first. It is a package var so a test can drive
+// the reboot/halt/power-off SystemAction verbs (to assert the #4108 F8 journal
+// wiring) WITHOUT actually taking the host down.
+var schedulePowerAction = func(systemctlArg string) {
+	go func() {
+		time.Sleep(1 * time.Second)
+		// context.Background(): a confirmed power action must not be
+		// cancelled by client disconnect. Errors ignored as before.
+		runTimeout(context.Background(), "systemctl", systemctlArg)
+	}()
+}
+
+func (s *Server) SystemAction(ctx context.Context, req *pb.SystemActionRequest) (*pb.SystemActionResponse, error) {
+	switch req.Action {
+	case "reboot":
+		slog.Warn("system reboot requested via gRPC")
+		// Journal BEFORE the box goes down: the fsynced record survives the
+		// reboot even though the slog.Warn journald line may not (#4108 F8).
+		s.logSystemAction("reboot")
+		schedulePowerAction("reboot")
+		return &pb.SystemActionResponse{Message: "System going down for reboot NOW!"}, nil
+
+	case "halt":
+		slog.Warn("system halt requested via gRPC")
+		s.logSystemAction("halt")
+		schedulePowerAction("halt")
+		return &pb.SystemActionResponse{Message: "System halting NOW!"}, nil
+
+	case "power-off":
+		slog.Warn("system power-off requested via gRPC")
+		s.logSystemAction("power-off")
+		schedulePowerAction("poweroff")
+		return &pb.SystemActionResponse{Message: "System powering off NOW!"}, nil
+
+	case "zeroize":
+		slog.Warn("system zeroize requested via gRPC")
+		// Journal BEFORE the wipe: the fsynced record is durable on disk
+		// before any config file is removed, so an interrupted wipe still
+		// leaves an attributable trail and a remote syslog collector keeps the
+		// cross-wipe record (#4108 F8). The wipe itself now removes
+		// .config.journal (a completed factory reset must not hand its audit
+		// log to the next tenant — #4576), superseding the earlier
+		// journal-survives-zeroize belt-and-braces.
+		s.logSystemAction("zeroize")
+		if err := performZeroizeWipe(); err != nil {
+			// A partial wipe may leave prior-tenant config/secrets on disk
+			// (#4576). Surface it instead of reporting a clean factory reset —
+			// the operator must know the erasure did not fully complete and
+			// the device is NOT safe to re-tenant.
+			slog.Error("system zeroize did not fully erase config state", "err", err)
+			return nil, status.Errorf(codes.Internal,
+				"zeroize incomplete: config state may remain on disk: %v", err)
+		}
+		return &pb.SystemActionResponse{Message: "System zeroized. Configuration erased. Reboot to complete factory reset."}, nil
+
+	case "clear-config-lock":
+		holder, locked := s.store.ConfigHolder()
+		if !locked {
+			return &pb.SystemActionResponse{Message: "No configuration lock held"}, nil
+		}
+		s.store.ForceExitConfigure()
+		return &pb.SystemActionResponse{Message: fmt.Sprintf("Configuration lock cleared (was held by %s)", holder)}, nil
+
+	case "clear-arp":
+		out, err := combinedOutputTimeout(ctx, "ip", "-4", "neigh", "flush", "all")
+		if err != nil {
+			return nil, status.Errorf(codes.Internal, "flush ARP: %s", strings.TrimSpace(string(out)))
+		}
+		return &pb.SystemActionResponse{Message: "ARP cache cleared"}, nil
+
+	case "clear-interfaces-statistics":
+		return &pb.SystemActionResponse{
+			Message: "Interface statistics counters noted\n(kernel counters are cumulative and cannot be reset)",
+		}, nil
+
+	case "clear-ipv6-neighbors":
+		out, err := combinedOutputTimeout(ctx, "ip", "-6", "neigh", "flush", "all")
+		if err != nil {
+			return nil, status.Errorf(codes.Internal, "flush IPv6 neighbors: %s", strings.TrimSpace(string(out)))
+		}
+		return &pb.SystemActionResponse{Message: "IPv6 neighbor cache cleared"}, nil
+
+	case "clear-policy-counters":
+		if s.dp == nil || !s.dp.IsLoaded() {
+			return nil, status.Error(codes.Unavailable, "dataplane not loaded")
+		}
+		if err := s.dp.ClearPolicyCounters(); err != nil {
+			return nil, status.Errorf(codes.Internal, "%v", err)
+		}
+		return &pb.SystemActionResponse{Message: "policy hit counters cleared"}, nil
+
+	case "clear-firewall-counters":
+		if s.dp == nil || !s.dp.IsLoaded() {
+			return nil, status.Error(codes.Unavailable, "dataplane not loaded")
+		}
+		if err := s.dp.ClearFilterCounters(); err != nil {
+			return nil, status.Errorf(codes.Internal, "%v", err)
+		}
+		return &pb.SystemActionResponse{Message: "Firewall filter counters cleared"}, nil
+
+	case "clear-nat-counters":
+		if s.dp == nil || !s.dp.IsLoaded() {
+			return nil, status.Error(codes.Unavailable, "dataplane not loaded")
+		}
+		if err := s.dp.ClearNATRuleCounters(); err != nil {
+			return nil, status.Errorf(codes.Internal, "%v", err)
+		}
+		return &pb.SystemActionResponse{Message: "NAT translation statistics cleared"}, nil
+
+	case "clear-persistent-nat":
+		if s.dp == nil || s.dp.GetPersistentNAT() == nil {
+			return &pb.SystemActionResponse{Message: "Persistent NAT table not available"}, nil
+		}
+		count := s.dp.GetPersistentNAT().Len()
+		s.dp.GetPersistentNAT().Clear()
+		return &pb.SystemActionResponse{
+			Message: fmt.Sprintf("Cleared %d persistent NAT bindings", count),
+		}, nil
+
+	case "ospf-clear":
+		if s.frr == nil {
+			return nil, status.Errorf(codes.FailedPrecondition, "FRR manager not available")
+		}
+		if _, err := s.frr.ExecVtysh("clear ip ospf process"); err != nil {
+			return nil, status.Errorf(codes.Internal, "clear OSPF: %v", err)
+		}
+		return &pb.SystemActionResponse{Message: "OSPF process cleared"}, nil
+
+	case "bgp-clear":
+		if s.frr == nil {
+			return nil, status.Errorf(codes.FailedPrecondition, "FRR manager not available")
+		}
+		if _, err := s.frr.ExecVtysh("clear bgp * soft"); err != nil {
+			return nil, status.Errorf(codes.Internal, "clear BGP: %v", err)
+		}
+		return &pb.SystemActionResponse{Message: "BGP sessions cleared (soft reset)"}, nil
+
+	case "ipsec-sa-clear":
+		if s.ipsec == nil {
+			return nil, status.Errorf(codes.FailedPrecondition, "IPsec manager not available")
+		}
+		count, err := s.ipsec.TerminateAllSAs()
+		if err != nil {
+			return nil, status.Errorf(codes.Internal, "%v", err)
+		}
+		return &pb.SystemActionResponse{
+			Message: fmt.Sprintf("Cleared %d IPsec SA(s)", count),
+		}, nil
+
+	case "dhcp-renew":
+		if s.dhcp == nil {
+			return nil, status.Errorf(codes.FailedPrecondition, "DHCP manager not available")
+		}
+		if req.Target == "" {
+			return nil, status.Errorf(codes.InvalidArgument, "dhcp-renew requires target interface")
+		}
+		if err := s.dhcp.Renew(req.Target); err != nil {
+			return nil, status.Errorf(codes.NotFound, "%v", err)
+		}
+		return &pb.SystemActionResponse{
+			Message: fmt.Sprintf("DHCP renewal initiated on %s", req.Target),
+		}, nil
+
+	case "in-service-upgrade":
+		if s.cluster == nil {
+			return nil, status.Error(codes.Unavailable, "cluster not configured")
+		}
+		if err := s.cluster.ForceSecondary(); err != nil {
+			return nil, status.Errorf(codes.FailedPrecondition, "ISSU: %v", err)
+		}
+		return &pb.SystemActionResponse{
+			Message: "Node is now secondary for all redundancy groups.\n" +
+				"Traffic has been drained to peer.\n" +
+				"You may now replace the binary and restart the service:\n" +
+				"  systemctl stop xpfd && <replace binary> && systemctl start xpfd",
+		}, nil
+
+	case "dynamic-dns-update", "dynamic-dns-check":
+		// #3276: operator force-now / check-now verb. force=true (update)
+		// re-asserts every owned DDNS record now; force=false (check) only
+		// re-observes + publishes changed records. The handler honors the
+		// per-RG owner gate (no-op + clear message on the backup).
+		if s.surfaceADDNSForceFn == nil {
+			return nil, status.Error(codes.Unavailable, "dynamic-dns: DDNS engine not running")
+		}
+		_, msg := s.surfaceADDNSForceFn(req.Action == "dynamic-dns-update")
+		return &pb.SystemActionResponse{Message: msg}, nil
+
+	default:
+		// Handle cluster failover actions: "cluster-failover:<rgID>" and "cluster-failover-reset:<rgID>"
+		if strings.HasPrefix(req.Action, "cluster-failover-reset:") {
+			if s.cluster == nil {
+				return nil, status.Error(codes.Unavailable, "cluster not configured")
+			}
+			rgStr := strings.TrimPrefix(req.Action, "cluster-failover-reset:")
+			rgID, err := strconv.Atoi(rgStr)
+			if err != nil {
+				return nil, status.Errorf(codes.InvalidArgument, "invalid redundancy-group ID: %s", rgStr)
+			}
+			if err := s.cluster.ResetFailover(rgID); err != nil {
+				return nil, status.Errorf(codes.NotFound, "%v", err)
+			}
+			return &pb.SystemActionResponse{
+				Message: fmt.Sprintf("Failover reset for redundancy group %d", rgID),
+			}, nil
+		}
+		if strings.HasPrefix(req.Action, "cluster-failover-data:node") {
+			if s.cluster == nil {
+				return nil, status.Error(codes.Unavailable, "cluster not configured")
+			}
+			nodeStr := strings.TrimPrefix(req.Action, "cluster-failover-data:node")
+			targetNode, err := strconv.Atoi(nodeStr)
+			if err != nil {
+				return nil, status.Errorf(codes.InvalidArgument, "invalid node ID: %s", nodeStr)
+			}
+			if !cluster.IsSupportedClusterNodeID(targetNode) {
+				return nil, status.Errorf(codes.InvalidArgument, "unsupported cluster failover target node %d", targetNode)
+			}
+			if targetNode != s.cluster.NodeID() {
+				if peerForwardedFromContext(ctx) {
+					return nil, status.Errorf(codes.FailedPrecondition, "forwarded cluster failover target node %d is not local", targetNode)
+				}
+				resp, err := s.proxyPeerSystemAction(ctx, req)
+				if err != nil {
+					if st, ok := status.FromError(err); ok {
+						return nil, st.Err()
+					}
+					return nil, status.Errorf(codes.Unavailable, "peer cluster failover proxy failed: %v", err)
+				}
+				return resp, nil
+			}
+
+			dataRGs := s.cluster.DataGroupIDs()
+			if len(dataRGs) == 0 {
+				return nil, status.Error(codes.FailedPrecondition, "no data redundancy groups configured")
+			}
+			moveRGs := make([]int, 0, len(dataRGs))
+			for _, rgID := range dataRGs {
+				if !s.cluster.IsLocalPrimary(rgID) {
+					moveRGs = append(moveRGs, rgID)
+				}
+			}
+			if len(moveRGs) == 0 {
+				return &pb.SystemActionResponse{
+					Message: fmt.Sprintf("All data redundancy groups are already primary on node %d", targetNode),
+				}, nil
+			}
+			if len(moveRGs) == 1 {
+				if err := s.cluster.RequestPeerFailover(moveRGs[0]); err != nil {
+					return nil, status.Errorf(codes.FailedPrecondition, "%v", err)
+				}
+			} else {
+				if err := s.cluster.RequestPeerFailoverBatch(moveRGs); err != nil {
+					return nil, status.Errorf(codes.FailedPrecondition, "%v", err)
+				}
+			}
+			return &pb.SystemActionResponse{
+				Message: fmt.Sprintf("Manual failover completed for data redundancy groups %v (transfer committed)", moveRGs),
+			}, nil
+		}
+		if strings.HasPrefix(req.Action, "cluster-failover:") {
+			if s.cluster == nil {
+				return nil, status.Error(codes.Unavailable, "cluster not configured")
+			}
+			rest := strings.TrimPrefix(req.Action, "cluster-failover:")
+
+			// Parse "cluster-failover:<rgID>[:node<N>]"
+			var rgStr, nodeStr string
+			if idx := strings.Index(rest, ":node"); idx >= 0 {
+				rgStr = rest[:idx]
+				nodeStr = rest[idx+5:] // skip ":node"
+			} else {
+				rgStr = rest
+			}
+
+			rgID, err := strconv.Atoi(rgStr)
+			if err != nil {
+				return nil, status.Errorf(codes.InvalidArgument, "invalid redundancy-group ID: %s", rgStr)
+			}
+
+			// If "node <N>" specified, route to correct node.
+			if nodeStr != "" {
+				targetNode, err := strconv.Atoi(nodeStr)
+				if err != nil {
+					return nil, status.Errorf(codes.InvalidArgument, "invalid node ID: %s", nodeStr)
+				}
+				if targetNode != s.cluster.NodeID() {
+					if peerForwardedFromContext(ctx) {
+						return nil, status.Errorf(codes.FailedPrecondition, "forwarded cluster failover target node %d is not local", targetNode)
+					}
+					resp, err := s.proxyPeerSystemAction(ctx, req)
+					if err != nil {
+						if st, ok := status.FromError(err); ok {
+							return nil, st.Err()
+						}
+						return nil, status.Errorf(codes.Unavailable, "peer cluster failover proxy failed: %v", err)
+					}
+					return resp, nil
+				}
+				// Target is local — make us primary.
+				if s.cluster.IsLocalPrimary(rgID) {
+					return &pb.SystemActionResponse{
+						Message: fmt.Sprintf("Redundancy group %d is already primary on node %d", rgID, targetNode),
+					}, nil
+				}
+				// Ask peer to transfer out so we can take primary.
+				if err := s.cluster.RequestPeerFailover(rgID); err != nil {
+					return nil, status.Errorf(codes.FailedPrecondition, "%v", err)
+				}
+				return &pb.SystemActionResponse{
+					Message: fmt.Sprintf("Manual failover completed for redundancy group %d (transfer committed)", rgID),
+				}, nil
+			}
+
+			if err := s.cluster.ManualFailover(rgID); err != nil {
+				return nil, status.Errorf(codes.NotFound, "%v", err)
+			}
+			return &pb.SystemActionResponse{
+				Message: fmt.Sprintf("Manual failover triggered for redundancy group %d", rgID),
+			}, nil
+		}
+		if strings.HasPrefix(req.Action, "userspace-inject:") {
+			provider, err := s.userspaceDataplaneControl()
+			if err != nil {
+				return nil, status.Error(codes.Unavailable, err.Error())
+			}
+			rest := strings.TrimPrefix(req.Action, "userspace-inject:")
+			parts := strings.SplitN(rest, ":", 2)
+			if len(parts) != 2 {
+				return nil, status.Error(codes.InvalidArgument, "usage: userspace-inject:<slot>:<mode>")
+			}
+			slot, err := strconv.Atoi(parts[0])
+			if err != nil {
+				return nil, status.Errorf(codes.InvalidArgument, "invalid userspace slot: %s", parts[0])
+			}
+			mode := parts[1]
+			statusNow, err := provider.Status()
+			if err != nil {
+				return nil, status.Errorf(codes.Unavailable, "userspace status: %v", err)
+			}
+			extra, err := dpuserspace.DecodeInjectPacketTarget(req.Target)
+			if err != nil {
+				return nil, status.Error(codes.InvalidArgument, err.Error())
+			}
+			injectReq, err := dpuserspace.BuildInjectPacketRequest(uint32(slot), mode, extra, statusNow)
+			if err != nil {
+				return nil, status.Error(codes.InvalidArgument, err.Error())
+			}
+			statusAfter, err := provider.InjectPacket(injectReq)
+			if err != nil {
+				return nil, status.Errorf(codes.FailedPrecondition, "userspace inject: %v", err)
+			}
+			msg := dpformat.FormatStatusSummary(statusAfter) + "\n" + dpformat.FormatBindings(statusAfter)
+			return &pb.SystemActionResponse{Message: msg}, nil
+		}
+		if strings.HasPrefix(req.Action, "userspace-forwarding:") {
+			provider, err := s.userspaceDataplaneControl()
+			if err != nil {
+				return nil, status.Error(codes.Unavailable, err.Error())
+			}
+			armed, err := dpuserspace.ParseForwardingCommand([]string{"forwarding", strings.TrimPrefix(req.Action, "userspace-forwarding:")})
+			if err != nil {
+				return nil, status.Error(codes.InvalidArgument, err.Error())
+			}
+			statusAfter, err := provider.SetForwardingArmed(armed)
+			if err != nil {
+				return nil, status.Errorf(codes.FailedPrecondition, "userspace forwarding control: %v", err)
+			}
+			msg := dpformat.FormatStatusSummary(statusAfter) + "\n" + dpformat.FormatBindings(statusAfter)
+			return &pb.SystemActionResponse{Message: msg}, nil
+		}
+		if strings.HasPrefix(req.Action, "userspace-queue:") {
+			provider, err := s.userspaceDataplaneControl()
+			if err != nil {
+				return nil, status.Error(codes.Unavailable, err.Error())
+			}
+			rest := strings.TrimPrefix(req.Action, "userspace-queue:")
+			parts := strings.SplitN(rest, ":", 2)
+			if len(parts) != 2 {
+				return nil, status.Error(codes.InvalidArgument, "usage: userspace-queue:<queue>:<register|unregister|arm|disarm>")
+			}
+			queueID, err := strconv.Atoi(parts[0])
+			if err != nil {
+				return nil, status.Errorf(codes.InvalidArgument, "invalid userspace queue: %s", parts[0])
+			}
+			registered, armed, err := dpuserspace.ParseRegistrationOperation(parts[1])
+			if err != nil {
+				return nil, status.Error(codes.InvalidArgument, err.Error())
+			}
+			statusAfter, err := provider.SetQueueState(uint32(queueID), registered, armed)
+			if err != nil {
+				return nil, status.Errorf(codes.FailedPrecondition, "userspace queue control: %v", err)
+			}
+			msg := dpformat.FormatStatusSummary(statusAfter) + "\n" + dpformat.FormatBindings(statusAfter)
+			return &pb.SystemActionResponse{Message: msg}, nil
+		}
+		if strings.HasPrefix(req.Action, "userspace-binding:") {
+			provider, err := s.userspaceDataplaneControl()
+			if err != nil {
+				return nil, status.Error(codes.Unavailable, err.Error())
+			}
+			rest := strings.TrimPrefix(req.Action, "userspace-binding:")
+			parts := strings.SplitN(rest, ":", 2)
+			if len(parts) != 2 {
+				return nil, status.Error(codes.InvalidArgument, "usage: userspace-binding:<slot>:<register|unregister|arm|disarm>")
+			}
+			slot, err := strconv.Atoi(parts[0])
+			if err != nil {
+				return nil, status.Errorf(codes.InvalidArgument, "invalid userspace slot: %s", parts[0])
+			}
+			registered, armed, err := dpuserspace.ParseRegistrationOperation(parts[1])
+			if err != nil {
+				return nil, status.Error(codes.InvalidArgument, err.Error())
+			}
+			statusAfter, err := provider.SetBindingState(uint32(slot), registered, armed)
+			if err != nil {
+				return nil, status.Errorf(codes.FailedPrecondition, "userspace binding control: %v", err)
+			}
+			msg := dpformat.FormatStatusSummary(statusAfter) + "\n" + dpformat.FormatBindings(statusAfter)
+			return &pb.SystemActionResponse{Message: msg}, nil
+		}
+		return nil, status.Errorf(codes.InvalidArgument, "unknown action: %s", req.Action)
+	}
+}
